% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WangXiaoToolsmyplot.R
\name{dplyrWrapper}
\alias{dplyrWrapper}
\title{Wrapper function for dplyer}
\usage{
dplyrWrapper(t)
}
\arguments{
\item{x}{data.frame}
}
\value{
dplyr
}
\description{
This is a wrapper for acquiring marginals and other slices of data
}
\examples{
data(d)
dplyrWrapper(d)
}
