% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WangXiaoToolsmyplot.R
\name{plotMyData1}
\alias{plotMyData1}
\title{2nd Wrapper function for ggplot2 for data d}
\usage{
plotMyData1(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
ggplot2
}
\description{
Computes the mean, variance and sd of a vector
}
\examples{
data(d)
plotMyData1(d)
}
