% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WangXiaoToolsmyplot.R
\name{plotMyData2}
\alias{plotMyData2}
\title{3rd Wrapper function for ggplot2 for data d}
\usage{
plotMyData2(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
ggplot2
}
\description{
Computes the mean, variance and sd of a vector
}
\examples{
data(d)
plotMyData2(d)
}
