% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t}
\alias{t}
\title{DATA3 for func3}
\format{An object of class \code{data.frame} with 5 rows and 5 columns.}
\usage{
t
}
\description{
DATA3 for func3
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p1.txt}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
